fixed a bug with the enemy applying force to the player
When the enemy collided with the player, he applied force to the player and the player would move without you trying to move in that direction. I added logic in the temp player movement script so that all forces applied to the player are reset. This works because the player is being moved by the transform in that script if you move the player by force you can fix this bug by increasing the players mass.