Fixed Wepon Bug
I made the characterBeingAttacked into a public variable so the whole script had access to it and attached the player prefab to it. I also cleaned up the code in the foreach so the character would take damage without being recreated every time through the loop.
(Recomendation: Change name from playerBeingAttacked to simply player or playerAttacked. The name suggests that there is more than one player to attack. Since there is not a name change is recomended to cut down on confusion)