finished the teleporting logic
finished the teleporting logic, you can walk through the whole game now!
Wrote a new camera movement script to follow the player.

added alt player starting position and logic.
because if you go from the hub to a facility you want to be at the start, but
if you are on the second level of a facility going back up you don't want to start
at the beginning you want to start at the top of the stairs.