Added doors and puzzle to hub
Added doors and puzzle to the hub.

Fixed door ways in the hub.

Added destructible Walls to the hub.

Added script that changes the players starting position in the hub so that you start at the top the first time but after you start at the bottom when you return to the hub.

Cleaned up code.