Finished the basilisk boss
Created a bomb prefab.

Added the basilisk to the biolab.

Changed the enable and disable sheild methods to be virtual in baseCharacter.
Finished and commented the basilisk script.

Changed the camera paning script to freeze player movement when the camera is panning.

Changed the player script to not take damage if the shield is up(because the shield pretty much useless and was doing nothing).

Added comments to the basilisk encounter manager.

Added a comment to the enemy spawner script.

Created a bomb script for the boss.