Finished implementing the gameplay for the biolab
Changed ranged enemy prefab (so that the bullet doesnt start in the enemies collider).
changed the ranged guardian prefab (so that the bullet doesnt start in the enemies collider).
Added gameplay logic to the biolab.
Changed bullet Logic script so that enemy bullets explode when they hit an enemy.
Changed doorLogic script to FIx a bug with a door opening at the wrong time if it was a door with a key.
Changed the enemySpawner script to not spawn in all enemies at once but one at a time every .5 seconds.
Changed the unlockDoorObject script, added that it will play a sound effect if there is a sound effect on that game object(lever).