Mostly added new logic to the enemy spawner script
Added a List of doors to lock/unlock to enemy spawner script. (doors unlock when THAT spawner’s enemies are defeated)

Added logic to make the camera pan optional.

Moved logic to freeze/unfreeze the player to be a method on the player script.

Added “Red points” for enemies and boss spawners.