Finished the shield guardain
Changed base character script to fix a bug with shoting in the wrong direction.
Changed player script to remove the shooting delay that caused problems with shooting in the wrong direction.
Added a bool to enemy spawner so that you have the option to freeze the players movement or not.