Added Kalebs health system to the game
Now the players health works if you start the game from the menu scene.

Changed so that the player is persistent across all scenes.

Changed so that the health system for the player is persistent across all scenes.

Hub changes (removed the player in that scene)

menu changes (added the player health canvas and the player)

game manager script (made the player and player health canvas persistent)

player script (fixed a bug that the health bar never changed)