Fixed heart increase
Fixed that the player can increase his number of “hearts”.

Removed not needed current health variable in base character. (we were storing the same data in two different places now its stored only in the float value.)

Removed not needed logic to not damage the player with fire attacks if he was behind a column.