lots of player health bar fixes
Added a bool on enemy’s that when true the enemy will always drop its power up game object.

Changed so that the player starts with 3 hearts.

Changed that the first light enemy that you kill in the hub will drop a heart increase. (so that you leave the hub with 4 hearts)

Changed the player prefab to have the players health canvas. (so that you have the players health bar even if you don’t start in “Start”)

Fixed that the full health pick up does not add an extra heart.

Fixed a bug where the players health was not displaying right.