player and health improvments
Fixed that the player and all enemy’s health’s lowest posable value is zero. (so the player’s health will not go to –22) (this fixes the issue that health pickups do basically nothing when the players health was to low)

Added faded rings on the players health bar when he does not have any power ups. (so that it doesn’t look so empty)