lots of changes read description for details
Merged implement gameplay into my branch.

Changed the health of the leviadrin to 10 and gave it its own float value.

Changed the enemies spawned in Atlantis to use more enemies than just the shield enemy.

Added health pickups in Atlantis. (no heart increase in this facility though)

Fixed the ranged guardians health bar to appear in the top right.

Changed that the ranged guardian drops the ranged weapon unlock, the basilisk drops the green shard, the shield guardian drops the shield unlock, and the Levaidrin drops the blue shard.

Added Brett’s new boss health bar to the ranged guardian, basilisk, shield guardian, Levaidrin, and Wyrm.

Changed the art for the key in the biolab to use an open source pixel art key.

Changed that when an enemy dies its colliders are disabled, so dead enemies won’t stop other enemies or the player from moving through them.

Added the ability for the player to “die” and go to a game over menu. (bool in Globals playerCanDie to turn this on and off)

Created a game over menu.

So now when the player “dies” the game goes into slow motion and the game over sound effect plays the scene fades to black and the game over menu scene loads.

Changed the weapon unlocks and shards sound effects to be controlled by the options menu volume slider.