Lots of bug fixes.
Fixed that the screen was black after returning to the game after dying.

Changed the hammer guardian to always drop the hammer unlock.

Changed the behemoth to always drop the yellow shard.

Fixed factory’s doors to not all be locked.

Added a heart increase and full health pickups in factory.

Fixes a bug where the player could trigger the first puzzle in the hub by stepping on it.

Changed the order of the teleporters in the hub to match the order there played in.

Fixed locked doors wrong size destructible walls, etc. In the space base.

Fixed the bug where if you set the volume to zero it played at full volume.