bug fixes
Removed one of the health increases in biolab because Luke fixed that Hubs health increase will only drop once.

Fixed a layering bug with the ranged guardians' bullets.

Fixed a Layering bug with the basilisk. (basilisk head under the door and background)

Fixed a Shadow layering issue in the hub.

Fixed that the player does not trigger the first hub puzzle by just stepping on it

Fixed the players health bar to not have “air gaps” in it anymore.

Fixed the bug where going back to the hub from another facility caused the player to be in total black. (the players starting position was not 0 not the Z axis causing the camera to move under the hubs tile set)