lots of bug fixes
Tested a bug where when returning to the hub you got stuck in the teleporter (I didn’t change anything, but it seems to be working perfectly now...)

I played through the whole game (except for geothermal) it took 26mins and two deaths.

Added a healing heart to just before the behemoth (I had to start the boss fight with 2 hearts twice before doing this, not ideal)

Added a heart to the hub so that you can get back to full health before starting the next facility.

Fixed a bug were the behemoth was not dropping the yellow shard when dying.

Fixed a layering bug were the players bullets went under conveyor belts.

Fixed a layering bug were levers in the 2nd level of space base where not visible.